+++ composer.json
@@ -7,19 +7,24 @@
         "services",
         "service-management"
     ],
-    "authors": [],
+    "authors": [{
+        "name": "Adrian Ocneanu",
+        "email": "aocneanu@gmail.com",
+        "homepage": "https://laravel-enso.com",
+        "role": "Developer"
+    }],
     "homepage": "https://github.com/laravel-enso/services",
     "type": "library",
     "license": "MIT",
     "require": {
         "php": ">=7.4.0",
         "laravel-enso/core": "4.8.*",
-        "laravel-enso/dynamic-methods": "1.1.*",
-        "laravel-enso/financials": "2.7.*",
-        "laravel-enso/forms": "3.4.*",
-        "laravel-enso/helpers": "1.15.*",
-        "laravel-enso/measurement-units": "1.1.*",
-        "laravel-enso/tables": "3.2.*"
+        "laravel-enso/dynamic-methods": "^2.0",
+        "laravel-enso/financials": "^3.0",
+        "laravel-enso/forms": "^4.0",
+        "laravel-enso/helpers": "^2.0",
+        "laravel-enso/measurement-units": "^2.0",
+        "laravel-enso/tables": "^4.0"
     },
     "autoload": {
         "psr-4": {
