+++ routes/api.php
@@ -2,8 +2,8 @@
 
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['web', 'auth', 'core'])
-    ->namespace('LaravelEnso\Services\App\Http\Controllers')
+Route::middleware(['api', 'auth', 'core'])
+    ->namespace('LaravelEnso\Services\Http\Controllers')
     ->prefix('api/services')
     ->as('services.')
     ->group(function () {
