+++ src/Forms/Templates/service.json
@@ -39,7 +39,7 @@
                     "value": null,
                     "meta": {
                         "type": "select",
-                        "options": "LaravelEnso\\Helpers\\App\\Enums\\VatRates"
+                        "options": "LaravelEnso\\Helpers\\Enums\\VatRates"
                     }
                 },
                 {
