+++ src/Http/Controllers/Options.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace LaravelEnso\Services\App\Http\Controllers;
+namespace LaravelEnso\Services\Http\Controllers;
 
 use Illuminate\Routing\Controller;
-use LaravelEnso\Select\App\Traits\OptionsBuilder;
-use LaravelEnso\Services\App\Models\Service;
+use LaravelEnso\Select\Traits\OptionsBuilder;
+use LaravelEnso\Services\Models\Service;
 
 class Options extends Controller
 {
