+++ src/Http/Requests/ValidateServiceRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace LaravelEnso\Services\App\Http\Requests;
+namespace LaravelEnso\Services\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
